finished ipload to bq script
