+++ 5_LT_companies_cars/src/upload_to_bigquery.py
@@ -7,12 +7,7 @@ from pydantic import BaseModel
 from typing import List
 
 # Parameters
-BUCKET_NAME = "lithuania_statistics"
-FOLDER_NAME = "companies_cars"
-FILE_NAME = "Atviri_JTP_parko_duomenys.csv"
-DATASET_NAME = "lithuania_statistics"
-TABLE_NAME = "companies_cars_raw"
-
+from mappings import file_configurations
 
 # --------------------------------- MODELS --------------------------------- #
 class TableSchema(BaseModel):
@@ -35,20 +30,60 @@ class UploadToBigQuery:
         self.config = config
         self.storage_client = storage.Client()
         self.bigquery_client = bigquery.Client()
+        self.transform_functions = {
+            "Atviri_JTP_parko_duomenys.csv": self.transform_data_companies,
+            "employees_salaries_raw.csv": self.transform_data_employees
+        }
 
     def read_file_from_gcs(self) -> pd.DataFrame:
-        """Reads a .csv file row by row and if error skips the row and logs the error."""
+        """Reads a .csv file row by row and if error skips the row, logs the error."""
         bucket = self.storage_client.bucket(self.config.bucket_name)
         blob = bucket.blob(f'{self.config.folder_name}/{self.config.file_name}')
-        blob_string = blob.download_as_string()
-        data_frame = pd.read_csv(io.BytesIO(blob_string))
-        return self.transform_data(data_frame)
+        blob_as_string = blob.download_as_string()
+        delimiter = self.get_delimiter()
+        data_frame = pd.read_csv(io.BytesIO(blob_as_string), sep=delimiter, on_bad_lines='skip', encoding='utf-8')
+        # Get the transform function based on the table name
+        transform_function = self.transform_functions[self.config.file_name]
+        data_frame = transform_function(data_frame)
+        return data_frame
+    
+    def get_delimiter(self) -> str:
+        """Returns a delimiter based on the file name."""
+        delimiters = {
+            "Atviri_JTP_parko_duomenys.csv": ",",
+            "employees_salaries_raw.csv": ";"
+        }
+        return delimiters.get(self.config.file_name, ",")
 
-    def transform_data(self, data_frame: pd.DataFrame) -> pd.DataFrame:
+    
+    # -------------------------------- TRANSFORMS -------------------------------- #
+    def transform_data_companies(self, data_frame: pd.DataFrame) -> pd.DataFrame:
         """Transforms the DataFrame by selecting columns and dropping missing values."""
         columns_to_keep = ['MARKE', 'KOMERCINIS_PAV', 'KATEGORIJA_KLASE', 'NUOSAVA_MASE', 'GALIA', 'GALIA_ELEKTR', 'DEGALAI', 'CO2_KIEKIS', 'CO2_KIEKIS__WLTP', 'TERSALU_LYGIS', 'GALIOS_MASES_SANT', 'PIRM_REG_DATA', 'PIRM_REG_DATA_LT', 'KODAS', 'PAVADINIMAS', 'SAVIVALDYBE', 'APSKRITIS']
         data_frame = data_frame[columns_to_keep]
         data_frame = data_frame.dropna(subset=['KOMERCINIS_PAV'])
+        return data_frame
+    
+    def transform_data_employees(self, data_frame: pd.DataFrame) -> pd.DataFrame:
+        """Transforms the DataFrame by selecting columns and dropping missing values."""
+        columns_to_keep = ['Juridinių asmenų registro kodas (jarCode)', 'Pavadinimas (name)', 'Savivaldybė, kurioje registruota(municipality)', 'Ekonominės veiklos rūšies kodas(ecoActCode)', 'Ekonominės veiklos rūšies pavadinimas(ecoActName)', 'Mėnuo (month)', 'Vidutinis darbo užmokestis (avgWage)', 'Apdraustųjų skaičius (numInsured)']
+        # Make a copy of the DataFrame
+        data_frame = data_frame[columns_to_keep].copy()
+        # convert month 202301 (int64) to 2023-01-01 (date)
+        data_frame['Mėnuo (month)'] = pd.to_datetime(data_frame['Mėnuo (month)'], format='%Y%m')
+        data_frame['periodas'] = data_frame['Mėnuo (month)'].dt.date
+        # drop month column
+        data_frame = data_frame.drop(columns=['Mėnuo (month)'])
+        # rename columns without lithuanian alphabet
+        data_frame = data_frame.rename(columns={
+            "Juridinių asmenų registro kodas (jarCode)": "kodas",
+            "Pavadinimas (name)": "pavadinimas",
+            "Savivaldybė, kurioje registruota(municipality)": "savivaldybe",
+            "Ekonominės veiklos rūšies kodas(ecoActCode)": "veiklos_kodas",
+            "Ekonominės veiklos rūšies pavadinimas(ecoActName)": "veiklos_pavadinimas",
+            "Vidutinis darbo užmokestis (avgWage)": "vidutinis_darbo_uzmokestis",
+            "Apdraustųjų skaičius (numInsured)": "apdraustuju_skaicius"
+        })
         return data_frame
 
     def upload_file_to_bigquery(self, data_frame: pd.DataFrame):
@@ -80,35 +115,26 @@ class UploadToBigQuery:
 
 # ----------------------------------- MAIN ----------------------------------- #
 if __name__ == "__main__":
-    schema_definitions = [
-        TableSchema(name='MARKE', data_type='STRING'),
-        TableSchema(name='KOMERCINIS_PAV', data_type='STRING'),
-        TableSchema(name='KATEGORIJA_KLASE', data_type='STRING'),
-        TableSchema(name='NUOSAVA_MASE', data_type='FLOAT'),
-        TableSchema(name='GALIA', data_type='FLOAT'),
-        TableSchema(name='GALIA_ELEKTR', data_type='FLOAT'),
-        TableSchema(name='DEGALAI', data_type='STRING'),
-        TableSchema(name='CO2_KIEKIS', data_type='INTEGER'),
-        TableSchema(name='CO2_KIEKIS__WLTP', data_type='INTEGER'),
-        TableSchema(name='TERSALU_LYGIS', data_type='STRING'),
-        TableSchema(name='GALIOS_MASES_SANT', data_type='FLOAT'),
-        TableSchema(name='PIRM_REG_DATA', data_type='DATE'),
-        TableSchema(name='PIRM_REG_DATA_LT', data_type='DATE'),
-        TableSchema(name='KODAS', data_type='STRING'),
-        TableSchema(name='PAVADINIMAS', data_type='STRING'),
-        TableSchema(name='SAVIVALDYBE', data_type='STRING'),
-        TableSchema(name='APSKRITIS', data_type='STRING'),
-    ]
+    for config_entry in file_configurations:
+        transform_function = None
+        
+        if config_entry["table_name"] == "Atviri_JTP_parko_duomenys":
+            transform_function = UploadToBigQuery.transform_data_companies
+        elif config_entry["table_name"] == "employees_salaries_raw":
+            transform_function = UploadToBigQuery.transform_data_employees
+        
+        if transform_function is None:
+            logging.error(f"Could not find transform function for {config_entry['table_name']}")
+            continue
 
-    config = UploadConfig(
-        bucket_name=BUCKET_NAME,
-        folder_name=FOLDER_NAME,
-        file_name=FILE_NAME,
-        dataset_name=DATASET_NAME,
-        table_name=TABLE_NAME,
-        table_schema=schema_definitions
-    )
-
-    uploader = UploadToBigQuery(config)
-    data_frame = uploader.read_file_from_gcs()
-    uploader.upload_file_to_bigquery(data_frame)
+        config = UploadConfig(
+            bucket_name="lithuania_statistics",
+            folder_name="companies_cars",
+            file_name=config_entry["file_name"],
+            dataset_name="lithuania_statistics",
+            table_name=config_entry["table_name"],
+            table_schema=config_entry["table_schema"]
+        )
+        uploader = UploadToBigQuery(config)
+        data_frame = uploader.read_file_from_gcs()
+        uploader.upload_file_to_bigquery(data_frame)
